# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Conarket",
        "description": "API RESTFul do projeto conarket",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/conab/admins": {
            "get": {
                "tags": [
                    "Administradores da Conab"
                ],
                "summary": "Retorna todos os administradores da Conab",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/User"
                                                    },
                                                    {
                                                        "properties": {
                                                            "phones": {
                                                                "type": "array",
                                                                "items": {
                                                                    "$ref": "#/components/schemas/Phone"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unathorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Administradores da Conab"
                ],
                "summary": "Atualiza os dados do administrador autenticado",
                "description": "Retorna os dados do administrador atualizado",
                "operationId": "update",
                "requestBody": {
                    "request": "Administrador",
                    "description": "Objeto de usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "schema": "UserRequest",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserUpdateRequest"
                                    },
                                    {
                                        "properties": {
                                            "phones": {
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/components/schemas/Phone"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "UserResponse",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "properties": {
                                                "phones": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Phone"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unathorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Administradores da Conab"
                ],
                "summary": "Registra um novo administrador",
                "description": "Retorna os dados do administrador registrado",
                "operationId": "store",
                "requestBody": {
                    "request": "Administradores",
                    "description": "Objeto de usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "schema": "UserRequest",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserStoreRequest"
                                    },
                                    {
                                        "properties": {
                                            "phones": {
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/components/schemas/Phone"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "UserResponse",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "properties": {
                                                "phones": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Phone"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unathorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/conab/admins/{userId}": {
            "get": {
                "tags": [
                    "Administradores da Conab"
                ],
                "summary": "Retorna um administrador pelo ID",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Id do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "UserResponse",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "properties": {
                                                "phones": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Phone"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unathorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Phone": {
                "properties": {
                    "number": {
                        "description": "Número do telefone no formato (99) 99999-9999",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "name": {
                        "description": "Nome do usuário",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email do usuário",
                        "type": "string"
                    },
                    "cpf": {
                        "description": "CPF do usuário no formato XXX.XXX.XXX-XX",
                        "type": "string"
                    },
                    "user_type": {
                        "description": "Tipo do usuário (ADMIN_COOP, ADMIN_CONAB ou CUSTOMER)",
                        "type": "string"
                    },
                    "profile_picture": {
                        "description": "URL da foto de perfil do usuário",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserStoreRequest": {
                "properties": {
                    "name": {
                        "description": "Nome do usuário",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email do usuário",
                        "type": "string"
                    },
                    "password": {
                        "description": "Senha do usuário",
                        "type": "string"
                    },
                    "cpf": {
                        "description": "CPF do usuário no formato XXX.XXX.XXX-XX",
                        "type": "string"
                    },
                    "profile_picture": {
                        "description": "URL da foto de perfil do usuário",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "properties": {
                    "name": {
                        "description": "Nome do usuário",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email do usuário",
                        "type": "string"
                    },
                    "password": {
                        "description": "Senha do usuário",
                        "type": "string"
                    },
                    "new_password": {
                        "description": "Nova senha do usuário",
                        "type": "string"
                    },
                    "cpf": {
                        "description": "CPF do usuário no formato XXX.XXX.XXX-XX",
                        "type": "string"
                    },
                    "profile_picture": {
                        "description": "URL da foto de perfil do usuário",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "description": "Token JWT",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Bearer <Token>"
            }
        }
    },
    "security": [
        []
    ]
}